infrastructure/tests/requirements.txt
boto3>=1.26.0
requests>=2.28.0
PyJWT[crypto]>=2.4.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-html>=3.1.0
pytest-cov>=4.0.0
coverage>=7.0.0